Use refactored modal in provider accept
